Follow the Google style guid instead of Sun's
